Working demo
The game runs in the app. No new function but there were issues loading the screen for some reason. Not sure why trying to display the screen once fixed it so i saved that code and commented it out for time being even though it still works being commented out now.